削除迄動作確認済み

+++ app/app.js
@@ -35,8 +35,12 @@ app.get('/api/v1/users/:id', (req, res) => {
   const id = req.params.id
 //``にする事でJSの構文を使うことができる 結果はrowに入ってくる
   db.get(`SELECT * FROM users WHERE id = ${id}`, (err, row) => {
-    // 取得した結果が返ってくる。
-    res.json(row)
+    if (!row) {
+      res.status(400).send({error:"Not Found"})
+    } else {
+      // 取得した結果が返ってくる。
+      res.status(200).json(row)
+    }
   })
   db.close()
 })
@@ -55,15 +59,13 @@ app.get('/api/v1/search', (req, res) => {
 })
 
 // 関数の共通化
-const run = async (sql, db, res, message) => {
+const run = async (sql, db) => {
   return new Promise((resolve, reject) => {
     db.run(sql, (err) => {
       // エラーが発生したときステータスを返す
       if (err) {
-        res.status(500).send(err) //SQL失敗→サーバーエラー
-        return reject();
+        return reject(err);
       } else {
-        res.json({message: message})
         return resolve()
       }
     })
@@ -73,21 +75,28 @@ const run = async (sql, db, res, message) => {
 //Create a new user
 // 今までapp.getでGETメソッドを作っていたが下記からPOSTメソッド
 app.post('/api/v1/users', async (req, res) => {
-  //Connect database
-  const db = new sqlite3.Database(dbPath)
-  // リクエストのbodyのnameは必須
-  const name = req.body.name
-  const profile = req.body.profile ? req.body.profile : "" //三項演算子
-  const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
+  if (!req.body.name || !req.body.name != "") {
+    res.status(400).send({error:"ユーザー名を入力してください"})
+  }else {
+    //Connect database
+    const db = new sqlite3.Database(dbPath)
+    // リクエストのbodyのnameは必須
+    const name = req.body.name
+    const profile = req.body.profile ? req.body.profile : "" //三項演算子
+    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
 
-  // 上記で設定している為resolve()かreject()が返ってくるまでrunの実行を待つ
-  await run(
-    `INSERT INTO users (name,profile,date_of_birth) VALUES ("${name}","${profile}","${dateOfBirth}")`,
-    db,
-    res,
-    "新規ユーザーを作成しました"
-  )
-  db.close()
+    try {
+      // 上記で設定している為resolve()かreject()が返ってくるまでrunの実行を待つ
+      await run(
+        `INSERT INTO users (name,profile,date_of_birth) VALUES ("${name}","${profile}","${dateOfBirth}")`,
+        db
+      )
+      res.status(201).send({message:"新規ユーザーを作成しました"})
+    } catch (e) {
+      res.status(500).send({error: e})
+    }
+    db.close()
+  }
 })
 
 //Update user data
@@ -96,20 +105,26 @@ app.put('/api/v1/users/:id', async (req, res) => {
   const db = new sqlite3.Database(dbPath)
   const id = req.params.id
 
-  // 現在のユーザー情報を取得する
-  db.get(`SELECT * FROM users WHERE id = ${id}`, async(err, row) => {
-    // 取得した結果が返ってくる。
-    const name = req.body.name ? req.body.name : row.name
-    const profile = req.body.profile ? req.body.profile : row.profile
-    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
-
-    await run(
-      `UPDATE users SET name="${name}",profile="${profile}",date_of_birth="${dateOfBirth}" WHERE id = ${id}`,
-      db,
-      res,
-      "ユーザー情報を更新しました"
-    )
-  })
+      // 現在のユーザー情報を取得する
+    db.get(`SELECT * FROM users WHERE id = ${id}`, async (err, row) => {
+      if (!row) {
+        res.status(400).send({ error: "指定されたユーザーは存在しません" })
+      } else {
+        // 取得した結果が返ってくる。
+        const name = req.body.name ? req.body.name : row.name
+        const profile = req.body.profile ? req.body.profile : row.profile
+        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
+        try {
+          await run(
+            `UPDATE users SET name="${name}",profile="${profile}",date_of_birth="${dateOfBirth}" WHERE id = ${id}`,
+            db
+          )
+          res.status(200).send({ message: "ユーザー情報を更新しました。" })
+        } catch (e) {
+          res.status(500).send({ error: e })
+        }
+      }
+    })
   db.close()
 })
 //Delete user data
@@ -117,13 +132,22 @@ app.delete('/api/v1/users/:id', async (req, res) => {
   //Connect database
   const db = new sqlite3.Database(dbPath)
   const id = req.params.id
-
-  await run(
-    `DELETE FROM users WHERE id = ${id}`,
-    db,
-    res,
-    "ユーザーを削除しました"
-  )
+  // ユーザーが存在するか確認してからデリートを実行→ただ、ＳＱＬ二回投げることになるからあまりよくない
+  db.get(`SELECT * FROM users WHERE id = ${id}`, async (err, row) => {
+    if (!row) {
+      res.status(400).send({ error: "指定されたユーザーは存在しません" })
+    } else {
+      try {
+        await run(
+          `DELETE FROM users WHERE id = ${id}`,
+          db
+        )
+        res.status(201).send({message:"ユーザーを削除しました"})
+      } catch (e) {
+        res.status(500).send({error:e})
+      }
+    }
+  })
   db.close()
 })
